data/principle-policy-corpus/requirements.txt
pyyaml>=6.0
tqdm>=4.65.0
beautifulsoup4>=4.12.2
requests>=2.31.0
